Refactor version field to use `smallint` across entities
Replaced `Integer` with `short` for the `version` field in Java entities and aligned corresponding database columns to use `smallint`. Improved efficiency and consistency by updating related database migration scripts and adjusting service logic where necessary.

+++ consumption/consumption-service/src/main/java/org/energycompany/entity/BaseEntity.java
@@ -57,7 +57,7 @@ public class BaseEntity {
             nullable = false,
             columnDefinition = "integer default 0"
     )
-    private Integer version;
+    private short version;
 
     @PrePersist
     public void prePersist() {
